+++ models/foodModel.js
@@ -31,10 +31,10 @@ const foodModel={
     updateFood: (foodId, updatedFoodData, callback) => {
         const query = 'UPDATE food SET ? WHERE foodid = ?';
         pool.query(query, [updatedFoodData, foodId], callback);
-    }
+    },
 
-    //food booking by user
-    ,getFoodDetails: (foodid, callback) => {
+    //FOOD BOOKING BY USER
+    getFoodDetails: (foodid, callback) => {
         const query = 'SELECT * FROM food WHERE foodid = ?';
         pool.query(query, [foodid], (error, results) => {
             if (error) {
@@ -46,6 +46,24 @@ const foodModel={
             // Return the first food found
             return callback(null, results[0]);
         });
+    },
+
+    bookFood: (bookingData, callback) => {
+        const query = 'INSERT INTO foodbooking SET ?';
+        pool.query(query, bookingData, callback);
+    },
+  
+    getUserDetails: (userid, callback) => {
+      const query = 'SELECT * FROM user WHERE userid = ?';
+      pool.query(query, [userid], (error, results) => {
+          if (error) {
+              return callback(error, null);
+          }
+          if (results.length === 0) {
+              return callback(null, null); // No user found with the given ID
+          }
+          return callback(null, results[0]); // Return the first user found
+      });
     }
